fix: 알림 서비스 수정 및 컴포넌트 위치 변경
- 이벤트 소스 연결을 따로 utils로 분리함
- Editor를 컴포넌트로 옮김
